\name{The_package-package}
\alias{The_package-package}
\alias{The_package}
\docType{package}
\title{
\packageTitle{The_package}
}
\description{
\packageDescription{The_package}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{The_package}
\packageIndices{The_package}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{The_package}

Maintainer: \packageMaintainer{The_package}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
